refactor: page indicator로 분리&재사용
